pyproject.toml
[tool.poetry]
name = "data_linter"
version = "0.1.0"
description = "data linter"
authors = ["Thomas Hirsch <thomas.hirsch@digital.justice.gov.uk>",
           "George Kelly <george.kelly@digital.justice.gov.uk>",
           "Tapan Perkins <tapan.perkins@digital.justice.gov.uk>",
           "Karik Isichei <karik.isichei@digital.justice.gov.uk>"]

[tool.poetry.dependencies]
python = "^3.7"
dataengineeringutils3 = "^1.0.1"
goodtables = "^2.5.0"
boto3 = "^1.14.7"
jsonschema = "^3.2.0"
pyyaml = "^5.3.1"
iam_builder = "^3.7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
moto = "^1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
data_linter = "data_linter.command_line:main"